hook abstracted and css changes
